refactor(activation): address PR #66 review comments
- Rewrite module docstring IFT equation in RST math (LaTeX) format
- Extract _lognormal_act to pyrcel/activation/_common.py; both _arg2000.py
  and _mbn2014.py now import from the shared module rather than one
  borrowing a private from the other
- Replace all "w.r.t." with "with respect to" throughout _mbn2014.py
