+++ pyrcel/activation/_arg2000.py
@@ -19,11 +19,11 @@ jax.config.update("jax_enable_x64", True)
 
 import jax.numpy as jnp  # noqa: E402
 from jax import Array  # noqa: E402
-from jax.scipy.special import erfc  # noqa: E402
 from jax.typing import ArrayLike  # noqa: E402
 
 from .. import constants as c  # noqa: E402
 from ..thermo import dv, dv_cont, es, ka_cont, sigma_w  # noqa: E402
+from ._common import _lognormal_act  # noqa: E402, F401
 
 
 def _kohler_crit_approx(T: ArrayLike, r_dry: ArrayLike, kappa: ArrayLike) -> tuple[Array, Array]:
@@ -40,34 +40,6 @@ def _kohler_crit_approx(T: ArrayLike, r_dry: ArrayLike, kappa: ArrayLike) -> tup
     r_crit = jnp.sqrt((3.0 * kappa * r_dry**3) / A)
     s_crit = jnp.sqrt((4.0 * A**3) / (27.0 * kappa * r_dry**3))
     return r_crit, s_crit
-
-
-def _lognormal_act(
-    smax: ArrayLike, sigma: ArrayLike, N: ArrayLike, sgi: ArrayLike
-) -> tuple[Array, Array]:
-    """Activated fraction of one lognormal mode.
-
-    Parameters
-    ----------
-    smax : float
-        Maximum parcel supersaturation (decimal).
-    sigma : float
-        Geometric standard deviation.
-    N : float
-        Total number concentration (any unit; returned N_act is in the same unit).
-    sgi : float
-        Modal critical supersaturation (decimal).
-
-    Returns
-    -------
-    N_act : float
-        Activated number concentration.
-    act_frac : float
-        Activated fraction (0–1).
-    """
-    ui = 2.0 * jnp.log(sgi / smax) / (3.0 * jnp.sqrt(2.0) * jnp.log(sigma))
-    N_act = 0.5 * N * erfc(ui)
-    return N_act, N_act / N
 
 
 def arg2000(
